tweaks after review
