--- scripts/getPackages.js
@@ -1,19 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-const fs = require('fs');
-const path = require('path');
-
-const PACKAGES_DIR = path.resolve(__dirname, '../packages');
-
-// Get absolute paths of all directories under packages/*
-module.exports = function getPackages() {
-  return fs
-    .readdirSync(PACKAGES_DIR)
-    .map(file => path.resolve(PACKAGES_DIR, file))
-    .filter(f => fs.lstatSync(path.resolve(f)).isDirectory());
-};
